Rename selectedCommit file
